Move AEAudioFormat.h to Utils where it better fits
