+++ xbmc/cores/AudioEngine/Engines/CoreAudio/CoreAudioAEStream.h
@@ -23,7 +23,7 @@
 #include <list>
 
 #include "ICoreAudioSource.h"
-#include "cores/AudioEngine/AEAudioFormat.h"
+#include "cores/AudioEngine/Utils/AEAudioFormat.h"
 #include "cores/AudioEngine/Interfaces/AEStream.h"
 #include "cores/AudioEngine/Utils/AEConvert.h"
 #include "cores/AudioEngine/Utils/AERemap.h"
