+++ xbmc/cores/AudioEngine/Engines/CoreAudio/ICoreAudioSource.h
@@ -21,7 +21,7 @@
 
 #include <AudioUnit/AudioUnit.h>
 
-#include "cores/AudioEngine/AEAudioFormat.h"
+#include "cores/AudioEngine/Utils/AEAudioFormat.h"
 #include "cores/AudioEngine/Interfaces/AE.h"
 #include "utils/StdString.h"
