+++ xbmc/cores/AudioEngine/Utils/AEUtil.h
@@ -19,7 +19,7 @@
  *
  */
 
-#include "../AEAudioFormat.h"
+#include "AEAudioFormat.h"
 #include "utils/StdString.h"
 #include "PlatformDefs.h"
 #include <math.h>
